+++ src/main/java/com/meongnyangerang/meongnyangerang/service/UserService.java
@@ -25,55 +25,8 @@ import org.springframework.transaction.annotation.Transactional;
 public class UserService {
 
   private final UserRepository userRepository;
-  private final AuthenticationCodeRepository authenticationCodeRepository;
-  private final MailComponent mailComponent;
 
-  // 인증 코드 발송
-  @Transactional
-  public void sendVerificationCode(String email) {
-
-    // 중복 가입 방지
-    if (userRepository.existsByEmail(email)) {
-      throw new MeongnyangerangException(DUPLICATE_EMAIL);
-    }
-
-    // 6자리 랜덤 숫자 생성(Math.random()은 예측 가능한 난수 생성 → 보안 취약 → SecureRandom 사용)
-    String code = String.format("%06d", new SecureRandom().nextInt(900_000));
-
-    // 기존 코드 삭제
-    authenticationCodeRepository.deleteAllByEmail(email);
-
-    authenticationCodeRepository.save(AuthenticationCode.builder()
-        .email(email)
-        .code(code)
-        .build());
-
-    // 이메일 발송
-    String subject = "회원가입 인증코드";
-    String text = "인증코드: " + code + "\n3분 안에 입력해주세요.";
-    mailComponent.sendMail(email, subject, text);
-  }
-
-  // 인증 코드 검증 (3분 이내인지 체크)
-  @Transactional
-  public void verifyCode(String email, String code) {
-    AuthenticationCode authCode = authenticationCodeRepository.findByEmail(email)
-        .orElseThrow(() -> new MeongnyangerangException(AUTH_CODE_NOT_FOUND));
-
-    // 3분 초과 체크
-    if (authCode.getCreatedAt().plusMinutes(3).isBefore(LocalDateTime.now())) {
-      throw new MeongnyangerangException(EXPIRED_AUTH_CODE);
-    }
-
-    // 코드 일치 여부 확인
-    if (!authCode.getCode().equals(code)) {
-      throw new MeongnyangerangException(INVALID_AUTH_CODE);
-    }
-
-    // 코드 사용 후 무효화 (삭제)
-    authenticationCodeRepository.delete(authCode);
-  }
-
+  // 사용자 회원가입
   public void registerUser(UserSignupRequest request) {
 
     // 중복 가입 확인
